# exec: vectorized merge join operator