--- a/pkg/sql/exec/colvec_tmpl.go
+++ b/pkg/sql/exec/colvec_tmpl.go
@@ -64,6 +64,28 @@ func (m *memColumn) Append(vec ColVec, colType types.T, toLength uint64, fromLen
 	}
 }
 
+func (m *memColumn) AppendSlice(
+	vec ColVec, colType types.T, destStartIdx uint64, srcStartIdx uint16, srcEndIdx uint16,
+) {
+	batchSize := srcEndIdx - srcStartIdx
+	outputLen := destStartIdx + uint64(batchSize)
+
+	switch colType {
+	// {{range .}}
+	case _TYPES_T:
+		if outputLen > uint64(len(m._TemplateType())) {
+			m.col = append(m._TemplateType()[:destStartIdx], vec._TemplateType()[srcStartIdx:srcEndIdx]...)
+		} else {
+			copy(m._TemplateType()[destStartIdx:], vec._TemplateType()[srcStartIdx:srcEndIdx])
+		}
+	// {{end}}
+	default:
+		panic(fmt.Sprintf("unhandled type %d", colType))
+	}
+
+	m.ExtendNulls(vec, destStartIdx, srcStartIdx, batchSize)
+}
+
 func (m *memColumn) AppendWithSel(
 	vec ColVec, sel []uint16, batchSize uint16, colType types.T, toLength uint64,
 ) {
@@ -93,12 +115,44 @@ func (m *memColumn) AppendWithSel(
 	}
 }
 
+func (m *memColumn) AppendSliceWithSel(
+	vec ColVec,
+	colType types.T,
+	destStartIdx uint64,
+	srcStartIdx uint16,
+	srcEndIdx uint16,
+	sel []uint16,
+) {
+	batchSize := srcEndIdx - srcStartIdx
+	switch colType {
+	// {{range .}}
+	case _TYPES_T:
+		toCol := append(m._TemplateType()[:destStartIdx], make([]_GOTYPE, batchSize)...)
+		fromCol := vec._TemplateType()
+
+		for i := 0; i < int(batchSize); i++ {
+			toCol[uint64(i)+destStartIdx] = fromCol[sel[i+int(srcStartIdx)]]
+		}
+
+		m.col = toCol
+	// {{end}}
+	default:
+		panic(fmt.Sprintf("unhandled type %d", colType))
+	}
+
+	m.ExtendNullsWithSel(vec, destStartIdx, srcStartIdx, batchSize, sel)
+}
+
 func (m *memColumn) Copy(src ColVec, srcStartIdx, srcEndIdx uint64, typ types.T) {
+	m.CopyAt(src, 0, srcStartIdx, srcEndIdx, typ)
+}
+
+func (m *memColumn) CopyAt(src ColVec, destStartIdx, srcStartIdx, srcEndIdx uint64, typ types.T) {
 	switch typ {
 	// {{range .}}
 	case _TYPES_T:
-		copy(m._TemplateType(), src._TemplateType()[srcStartIdx:srcEndIdx])
-		// {{end}}
+		copy(m._TemplateType()[destStartIdx:], src._TemplateType()[srcStartIdx:srcEndIdx])
+	// {{end}}
 	default:
 		panic(fmt.Sprintf("unhandled type %d", typ))
 	}
@@ -225,5 +279,39 @@ func (m *memColumn) PrettyValueAt(colIdx uint16, colType types.T) string {
 	// {{end}}
 	default:
 		panic(fmt.Sprintf("unhandled type %d", colType))
+	}
+}
+
+func (m *memColumn) ExtendNulls(
+	vec ColVec, destStartIdx uint64, srcStartIdx uint16, toAppend uint16,
+) {
+	outputLen := destStartIdx + uint64(toAppend)
+	if uint64(cap(m.nulls)) < outputLen/64 {
+		// (batchSize-1)>>6+1 is the number of Int64s needed to encode the additional elements/nulls in the ColVec.
+		// This is equivalent to ceil(batchSize/64).
+		m.nulls = append(m.nulls, make([]int64, (toAppend-1)>>6+1)...)
+	}
+	if vec.HasNulls() {
+		for i := uint16(0); i < toAppend; i++ {
+			if vec.NullAt(srcStartIdx + i) {
+				m.SetNull64(destStartIdx + uint64(i))
+			}
+		}
+	}
+}
+
+func (m *memColumn) ExtendNullsWithSel(
+	vec ColVec, destStartIdx uint64, srcStartIdx uint16, toAppend uint16, sel []uint16,
+) {
+	outputLen := destStartIdx + uint64(toAppend)
+	if uint64(cap(m.nulls)) < outputLen/64 {
+		// (batchSize-1)>>6+1 is the number of Int64s needed to encode the additional elements/nulls in the ColVec.
+		// This is equivalent to ceil(batchSize/64).
+		m.nulls = append(m.nulls, make([]int64, (toAppend-1)>>6+1)...)
+	}
+	for i := uint16(0); i < toAppend; i++ {
+		if vec.NullAt(sel[srcStartIdx+i]) {
+			m.SetNull64(destStartIdx + uint64(i))
+		}
 	}
 }