mergeJoinOp is an operator that implements sort-merge join.
It performs a merge on the left and right input sources, based on the equality
columns, assuming both inputs are in sorted order.

The merge join operator uses a "two scan" approach to generate the join.
What this means is that instead of going through and expanding the cross
product row by row, the operator performs two passes.
The first pass generates a list of groups of matching rows based on the
equality column. A group is an ADT representing a run, or in other words,
multiple repeated values. This run is represented as the starting row ordinal,
the ending row ordinal, and the number of times this group is
expanded/repeated.
The second pass is where each of these groups are either expanded or repeated
into the output or savedOutput buffer, saving big on the type introspection.

Two buffers are used, one for the run on the left table and one for the run on
the right table. These buffers are only used if the run ends with a batch, to
make sure that we don't miss any cross product entries while expanding the
groups (leftGroups and rightGroups) when a run spans multiple batches.

Some preliminary benchmarking results:
```
BenchmarkMergeJoiner/rows=1024-8                      	   20000	     66913 ns/op	 979.42 MB/s
BenchmarkMergeJoiner/rows=4096-8                      	    5000	    265602 ns/op	 986.98 MB/s
BenchmarkMergeJoiner/rows=16384-8                     	    2000	   1042486 ns/op	1005.84 MB/s
BenchmarkMergeJoiner/rows=1048576-8                   	      20	  65653445 ns/op	1022.17 MB/s

BenchmarkMergeJoiner/oneSideRepeat-rows=1024-8         	   20000	     68281 ns/op	 959.79 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=4096-8         	    5000	    267793 ns/op	 978.90 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=16384-8        	    2000	   1049688 ns/op	 998.94 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=1048576-8      	      20	  67160497 ns/op	 999.23 MB/s

BenchmarkMergeJoiner/bothSidesRepeat-rows=1024-8       	   20000	     69234 ns/op	 946.58 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=4096-8       	    2000	    605505 ns/op	 432.93 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=16384-8      	     500	   2841285 ns/op	 369.05 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=1048576-8    	       3	 428893269 ns/op	 156.47 MB/s
```

Comparing this to the previous iteration
```
BenchmarkMergeJoiner/rows=1024-8                      	    5000	    246388 ns/op	 265.99 MB/s
BenchmarkMergeJoiner/rows=4096-8                       	    2000	    838724 ns/op	 312.55 MB/s
BenchmarkMergeJoiner/rows=16384-8                      	     500	   3105394 ns/op	 337.66 MB/s
BenchmarkMergeJoiner/rows=1048576-8                    	      10	 202760327 ns/op	 330.98 MB/s

BenchmarkMergeJoiner/oneSideRepeat-rows=1024-8         	    5000	    286032 ns/op	 229.12 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=4096-8         	    2000	    856202 ns/op	 306.17 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=16384-8        	     500	   3128651 ns/op	 335.15 MB/s
BenchmarkMergeJoiner/oneSideRepeat-rows=1048576-8      	      10	 194073135 ns/op	 345.79 MB/s

BenchmarkMergeJoiner/bothSidesRepeat-rows=1024-8       	    5000	    329374 ns/op	 198.97 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=4096-8       	    3000	    463981 ns/op	 564.99 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=16384-8      	    1000	   1190897 ns/op	 880.49 MB/s
BenchmarkMergeJoiner/bothSidesRepeat-rows=1048576-8							DNF (test was broken when I ran this ¯\_(ツ)_/¯)
```

Note that V2 outperforms V1 in every case except some of the degenerate `bothSidesRepeat` cases with a lot of rows. In this case, the cross product blows up and spends a bunch of time allocating memory (which didn't happen in V1 due to the row by row approach).

Release note: None