@georgeutsin 2019-02-19
Note: pushed a cherry pick to use `chunkBatchSource`, but ended up putting the test in the same commit. I will be sure to fix this later once the hash aggregator PR is merged in.