@jordanlewis 2019-02-20
I see. I think we should just go whole hog on this and fix the template
functions to rename their input parameters to the formal parameters
declared in the Go function. Should be fun!

On Wed, Feb 20, 2019, 10:21 George Utsin <notifications@github.com> wrote:

> *@georgeutsin* commented on this pull request.
>
> *Reviewable <https://reviewable.io/reviews/cockroachdb/cockroach/34676>*
> status: [image: :shipit:] complete! 0 of 0 LGTMs obtained (waiting on
> @georgeutsin <https://github.com/georgeutsin>, @jordanlewis
> <https://github.com/jordanlewis>, and @yuzefovich
> <https://github.com/yuzefovich>)
>
> ------------------------------
>
> *pkg/sql/exec/mergejoiner_tmpl.go, line 179 at r7
> <https://reviewable.io/reviews/cockroachdb/cockroach/34676#-LZANhD53np1mF4WZ7bS:-LZATluF8HnK8NuRO7hH:b118woi>
> (raw file
> <https://github.com/cockroachdb/cockroach/blob/6350533c83ce6d0521000e48dd000a6f578c49da/pkg/sql/exec/mergejoiner_tmpl.go#L179>):*
> *Previously, jordanlewis (Jordan Lewis) wrote…*
>
> Not sure why you have to copy all of the variables like this above? I
> would drop the t_ prefix from the template copy and make the names line up
> correctly rather than introduce all the t_ variables just for cleanliness.
> Maybe you needed to do this though?
>
> I overload certain words to mean different things depending on the
> context, like destStartIdx, as destination could either mean the first
> empty space in the ColVec(before the build call is made), or the current
> row to write into (within the function).
> This was supposed to emulate making a function call by scoping the context
> of the overloaded variables.
>
> I suppose I could rename the macro/micro version to different names but
> this would sacrifice consistency, at least partially
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/cockroachdb/cockroach/pull/34676#pullrequestreview-205819411>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AACrLXzATKeWXN1gRgzLQu-3lpZpOetMks5vPWfwgaJpZM4amCvH>
> .
>