@georgeutsin 2019-02-26
Found a bug hidden in a TODO that supposed to be a sanity check. Added test cases :sob:
![screen shot 2019-02-26 at 3 34 10 pm](https://user-images.githubusercontent.com/5915070/53444379-fe628000-39db-11e9-9ecb-97aa52dc8d0c.png)

Ready for that 👀 @cockroachdb/sql-execution